+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - make abs_url a request attribute
 - use prepared testbase image for pipeline testing
 - add environment variable substitution in site.ini
+- check proxy when determining protocol
+- make python 3.9 default for zoom bin command
 
 ## [6.25.1] - 2025-01-17
 - fix escape sequence deprecation warning in fill.py